Add new blockchain hosts extra var: peering_network
Previously the network interface to use to get each blockchain host in ansible was hard-coded (eth1).
The issue is that it only works for vagrant as network interfaces names can be different for each provider.

Those network interfaces names are used to get each bootnodes IPs.

Wrong interfaces lead to crash during deployment.

Now we try to get the network interface from the network role.